Creating file
This script replaces a character within all phone number fields for ever user in Active Directory. For example, if your company wants to move from storing phone numbers as "123-456-7890" to "123.456.7890", change the $characterToReplace to '-' and the $newCharacter to '.'.

When the script prompts, enter a username/password with permissions to change AD user attributes values.